requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
bcrypt==5.0.0
beanie==1.29.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorgram.py==1.2.0
cryptography==46.0.3
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.8.0
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.121.2
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
instaloader==4.15
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==1.0.7
langchain-core==1.0.5
langchain-tavily==0.2.13
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.4.43
lazy-model==0.2.0
llvmlite==0.45.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.21.1
mdurl==0.1.2
more-itertools==10.8.0
motor==3.7.1
moviepy==2.2.1
mpmath==1.3.0
multidict==6.7.0
networkx==3.4.2
nltk==3.9.2
numba==0.62.1
numpy==2.2.6
openai-whisper==20250625
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pillow==11.3.0
pip==21.2.3
proglog==0.1.12
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydantic-settings==2.12.0
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.4
pyparsing==3.2.5
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.29.0
rsa==4.9.1
setuptools==57.4.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.4
sse-starlette==3.0.3
starlette==0.49.3
sympy==1.14.0
tavily==1.1.0
tenacity==9.1.2
textblob==0.19.0
tiktoken==0.12.0
toml==0.10.2
torch==2.9.1
tqdm==4.67.1
typer==0.20.0
typing_extensions==4.15.0
typing-inspection==0.4.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
xxhash==3.6.0
yarl==1.22.0
youtube-transcript-api==1.2.3
yt-dlp==2025.11.12
zstandard==0.25.0
